Improve convergence info and include device types
Summary:
I've tried to improve the convergence section of the HIG. Most of it is
based on this wiki article:
https://community.kde.org/Plasma/Convergence_Overview

It now includes a list of device types and identifies common UI
components, as well as an entire page describing the optimal UX for
each device type. This should describe the big picture of convergence,
not just looking at single applications but how it effects the entire
system. This should help place the discussions on convergence for each
part of the whole in the proper context. And it should help align those
parts in a consistent way.

The Device Type UX page is strongly related to the rewrite of the
convergence section which is why I made it part of the same commit.

See also: T8146, T2175

Test Plan: Ran "make html" locally to verify the rewritten pages.

Reviewers: fabianr, #kde_human_interface_guidelines, colomar, ngraham

Reviewed By: fabianr, #kde_human_interface_guidelines, colomar, ngraham

Subscribers: colomar, ngraham

Tags: #kde_human_interface_guidelines

Maniphest Tasks: T9701

Differential Revision: https://phabricator.kde.org/D15385