Fix incorrect code in using_actions.md.
It is said that no changes to `main.cpp` are required, but actually since we are going to install UI rc file under the component `texteditor`, and `main.cpp` says `mainwindow`, the final build will fail to find the rc file.

Thanks to @swaptr on Matrix for pointing this out.