Mention both desktoptheme folders like we do in plasma/_index.md
We duplicate the two folders again next to the theme details since
it's more likely to be linked. We need to reorder the sections so that
the svg list is after the metadata section.

Wrap folder structure in code tags. strong > code uses the default
code CSS but still looks better.