Update docs for ServiceMenu to no longer reference deprecated install location
With this change being in KDE Frameworks 5.85, it should be broadly available by now.

Also, this commit fixes a typo of stating that an entry was added to the ServiceTypes, but
actually it was added to the MimeType. While this works, it is confusing and sth. that is
being phased out.

Relates to https://phabricator.kde.org/T14543