Simplify merge request templates
The current one is too annoying and makes it harder to contribute small
fixes from the gitlab webinterface. One of the original goal of
this website was to make it as easy to contribute to it as a wiki.

This results in many MR to not follow the template anyway.

The "how to test" also assume that some gitlab tools are installed
and this is only helpful for maintainer who anyway know how to test a
MR.