Fix obvious PlasmaCore -> Kirigami issues
This fixes the links from PlasmaCore.Units by pointing them to Kirigami.Units as per https://invent.kde.org/plasma/plasma-workspace/-/issues/74.

This MR is just for fixing the current pipeline failures, so I only changed superficial things.

It is not a rewrite or main fix of the tutorial, so don't review it as such. That would come in a future MR.