Update for KCM tutorial
KCM tutorial was a bit outdated since there was some Qt5 related things, I changed it a bit to get it to compile on Qt6 system.

Most changes were in the CMakeFiles so I added another in the root of the kcm folder, then second in src/, like is seen often in other projects. This should hopefully help people understand why there's many of them, and help people to figure out how to add libraries etc. to their KCM.

I also modified the structure a bit to something I often see.