ide/vscode: replace manual config section with template location note
The manual configuration section was overly detailed and not necessary,
making the guide harder to follow. Great care has been taken with the
automatic configuration in kdesrc-build to ensure that it makes getting
started with Visual Studio Code in KDE development as easy and automatic
as possible.

A `Notes` section has been added to the end of the guide to provide
information on where the template files are located in the kdesrc-build
repository, which can be referenced if the user wants to manually
configure Visual Studio Code without making the guide needlessly
complicated.