Kirigami Getting Started: Revert to using ~/.local install prefix
This reverts part of the changes from
https://invent.kde.org/documentation/develop-kde-org/-/merge_requests/404

The install prefix ~/.local was used because it conforms to the XDG
base directory specification. ~/.local/bin is expected to be a part
of PATH and ~/.local/share/applications is expected to be in XDG_DATA_DIRS,
which means the installed application will show up on the launcher menu
and the binary will show up on the terminal by default without the user
manually setting anything. With the new declarative registration,
QML_IMPORT_PATH doesn't need to be set up either. It just works.

Setting ~/kde/usr as the prefix breaks this for users of raw CMake
and requires sourcing the generated prefix file.

This commit also makes it clear that these CMake instructions
are intended only for those not using kdesrc-build/kde-builder.