HIG/Icons: Add more detail about how to choose the right icon
Right now the page is a bit light on recommendations regarding how to
choose which icons to use. This commit fixes that by implementing the
changes proposed and agreed to in
https://invent.kde.org/teams/vdg/issues/-/issues/55 regarding how to
choose universal vs specific icons. I've written a new section about
it, complete with images and examples.

I tried to avoid turning this into a "change everything" MR, but failed
as all the content on the page is fairly inter-related. As such,
changes beyond the new "universal or specific icons?" section include:

- Remove the "General icon use guidelines" section because all that
  information is now moved into specific sections.
- Re-arrange sections to flow more logically, including moving the
  "implementation details" section to the end.
- Make small re-wordings in various sections to make them read more
  nicely given their new content and positions.
- Don't call a black trash can icon red; clarify that it's only related
  to destructiveness.

Closes https://invent.kde.org/teams/vdg/issues/-/issues/55