HIG/text_and_labels.md: Change "subject" to "topic"
The HIG currently refer to the word "User" in "Add New User" as a
subject. This is add odds with school grammar and general linguistic
terminology, where "subject" (for English) generally refers to the
noun (phrase) before the verb (already omitted in the example due
to being in the imperative mood), in a standard active sentence
typically corresponding to the person or thing that performs an
action, while the term "object" refers to the noun (phrase) after
the verb, in a standard active sentence typically the person or
thing that is acted upon.