hig/text_and_labels: no ellipses on menu buttons
The HIG currently do not explicitly address menu buttons. These often
have action-based labels like "Add" or "Sort by", so if these are seen
as buttons, the guidance on adding ellipses should apply; but if they
are considered as menu categories, no ellipses are necessary (like the
"Edit" or "View" menu categories, which are actions that pop up a menu,
but should not have labels). Most applications currently take the second
perspective and do not have the ellipsis, but some have them. This
should be made consistent through explicit guidance.

This change amends the HIG to explicitly recommend the use of downward-
pointing arrows, already commonly used to indicate menu buttons, instead
of ellipses.