Create a web worker to create the lunr search index.
This reduces the 500ms-1500ms page "jank" per page load. If the user's
browser doesn't support web workers it should still run like it use to.

This does add the equivalent loading delay, and a "wait" cursor is
placed over the disabled search input. Once the data is retrieved from
the worker it is re-enabled.

This also replaces the remote lunr dependency with a local one.