Add a small button before each code block that copies to clipboard.
This change adds a new javascript file `copy-to-clipboard.js` that runs
on window `load` event that places a button before each `.highlight`
code block. Additionally an HTML template is used to assist in string
extraction for localization.

Fixes #167 

---

I'd like some advice on the placement of the button. It didn't look particularly nice over the code, but it's also possible that the icon itself is too small. If we don't mind the extra bar on top of every `<code>` element, we could stick in some meta-data about the block like the language used.

I've also added a small reaction icon for about 1.5 seconds after you press it to indicate success. Hopefully the icon I selected is approriate there!

I wrote it in plain javascript because I'm also working on a bootstrap 5 port which removes JQuery. So future proofing this :smile:. And I used an HTML template element to assist string extraction, let me know if you folks would rather have it be generated in javascript though.

Screenshot:

![image](/uploads/f7d47a2308f91363bbe61a54b9b7de42/image.png){width=599 height=544}
![image](/uploads/9b11550dbdc59ef27ee9653b39ec63ba/image.png){width=603 height=848}

Video:

![Screencast_20240808_222247](/uploads/26206e461fdfd89c59adf06d66054c8b/Screencast_20240808_222247.webm)