Add PySide6 Tutorial
This MR is a result of documentation/develop-kde-org#148, where once a Flatpak BaseApp for PySide6 is available, then the documentaiton can be updated to also include PySide6. The Flatpak BaseApp for PySide6 is now available.

As such, since many of the things (especially Packaging and QML files) are the same, a new tutorial was not created. Instead the current one is updated to also include PySide6 code sections. One can switch using tabs between the PySide6 and PyQt6 code sections. Some documentation was also updated to explicitly point out the difference between PySide6 and PyQt6.

Closes #148