hig/text_and_labels: recommend sentence case for form labels
The HIG contain a list of cases where sentence case is appropriate. This
list contains the labels for a few types of controls, namely radio
buttons, checkboxes, and comboboxes, but it does not include labels
placed in front of controls generally. A literal reading of the current
text would be that e.g. a Kirigami.FormData.label for a checkbox should
be sentence case, but one for a slider should be title case. This is not
how these labels are generally used currently, and would lead to a very
inconsistent look. Another case that is currently not addessed would be
the label for the star rating filter in Elisa's NavigationActionBar.

This change adds labels placed in front of a control to the list of
contexts where sentence case should be used.