Fix the Bootstrap regression that broke tab switching.
This adds the missing `-bs-` on `data-toggle` which is required in
Bootstrap 5. This change also tidies up some hard-coded values in the
tab-qt shortcode view, along with adding a "defaultTab" parameter.

This defaultTab parameter is required for at least one of the tabs in a
tabset-qt and makes code visible by default. (This does not affect the
default selected tab, just the tab's content.)

Finally this change removes the fade animation which was a bit excessive
in testing.

---

Fixes #178 

The removal of the fade animation is a bit subjective, but was a bit much imo. So let me know if you'd like me to revert that part. ~~I did a small bit of research but I couldn't find a nice way to pass "Hey you're the selected tab" to the tab-qt shortcode without completely refactoring the shortcode. So if anyone has any suggestions there feel free to relay them!~~

This now uses an Ordinal to determine nested shortcode position. Additionally it supports unlimited number of tabs, and spaces in tab names.