Adjust the tutorials to no longer include the missing files.
This also swaps some comment arrows to use the emphasis param for
highlighting.

The tutorial will need to be double checked.

---

Fixes #177 

To the best of my knowledge I've adjusted the tutorial, however my knowledge of qml/qt has faded since I last worked on a QT project. I'm sure my language here can be adjusted.

The main purpose of this MR is to adjust the extract plasma script to download the correct files, and write to stderr if there's a non-200 return code (and then skip the file.) 

I've adjusted/removed the files that 404'd. The `examples/all.cpp` went unused afaik.