Set up the groundwork to make the Kirigami tutorial multilingual
This is an introductory organization change to prepare the Kirigami tutorial to support Python and Rust.

This:

* Pushes the project setup to its own section "setup" so that we can have a single page for C++, another for Python, and another for Rust.
* Creates a new section "data" that is specifically for exposing data from C++/Python/Rust to QML.
* Because of the new "data" section, the old "advanced" section can then be used to explain complicated topics about the project setup and other complex topics.
* The index page is made more neutral to accomodate for multiple languages later on.
* Certain very C++ pages are made explicitly about using C++.

This is deliberately incomplete. It is to be followed up by subsequent MRs for Python and Rust,
where the true redesign thought will actually come up as it's written. Some parts will be rewritten.

By itself, this does not change the contents of the tutorial.