Replace old tab/tabs shortcodes with tab/tabset
What this does is remove the old tab and tabs shortcodes (which I tested and don't work very well)
with the new tabset-qt and tab-qt, while also renaming the latter to just tabset and tab.

The reason for this is that the new version is actually language agnostic and you have more control
about the code that will be displayed.

The old tab and tabs shortcodes only really allowed for Python and C++, which in general could work,
but then we'd just be extending that shortcode over and over whenever we need more langs.

We also have the code-toggle shortcode, but that only really allows JSON and YAML, so it suffers from
the same problems. That one however can be dealt with in another MR.