Add back kdesrc-build alerts
This was not meant to be removed, and I missed this when reviewing https://invent.kde.org/documentation/develop-kde-org/-/merge_requests/583.

The alert both grabs the attention of the user to say "this isn't for me" in case it isn't and it makes the tutorial shorter, showing only information that is relevant to them.