plasma/widget/_index.md the command lines did not work
Because the tutorial was ported from a time when metadata.json
was named metadata.desktop. The sed command lines are incorrect.
The sed command lines work correctly for metadata.desktop.
The sed command lines do not work correctly for metadata.json file.
In json files the last element in an object or list should not have
a trailing comma.