Add list of Python bindings packages
This uses installpackage to list all currently available bindings
on openSUSE, Fedora and Arch.

Arch was the first to have them, Fedora just added them and
I'm looking into getting them on openSUSE.

This MR is blocked until openSUSE adds them. It might be subject to change in the meantime. If openSUSE takes too long, we can delay adding them to the list and merge this.