HIG/Text and labels: improve clarity of icon style guide
Right now there's a bunch of text, and some of it is a bit wishy-washy.
Use a table to make the recommendation clearer, and use some of the
saved space to remind the reader that the whole symbolic vs colorful
thing is an icon theme decision anyway, and all you can do is ask it
nicely.