Fix /docs frontapage has wrong layout
Setting the layout to home does for some reason not work anymore. The home layout is usually meant to be automatically applied to pages of kind "home", but /docs is a sub dir and hence has page "section". Rename the layout to a non-reserved name (frontpage) to fix this.

Note: setting the kind: home in frontmatter is not a good option since it is deprecated (https://github.com/gohugoio/hugo/issues/12484)

Fixes #238