HIG/Getting Input: clarify some ambiguous situations with basic controls
There are some cases where it may not be clear when to use a specific
type of basic control, and when to use a different one.

This page already has some explanation for how to handle a few of these
situations, but it's all very condensed into a dense bulleted list,
which may be a bit intimidating and isn't super scannable.

Instead, break out the common cases into their own sections with shorter
and clearer explanations, plus a few graphics.

As a consequence, the page is now significantly longer and less dense.